+++ lib/e_way_client/factories/province.rb
@@ -4,7 +4,7 @@ FactoryBot.define do
     sequence(:province_id) {|n| n}
     sequence(:province_name_eng) {|n| "Eng province #{n}"}
     sequence(:province_name_viet) {|n| "Viet province #{n}"}
-    response_code "0000"
+    response_code { "0000" }
   end
 
 end