+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Factories use FactoryBot instead of FactoryGirl
+
 ## [1.2.1] - 2017-04-13
 ### Fixed
 - Query list provinces