+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2017-12-28
+### Added
+- Drop support FactoryBot less than 5.x
+
 ## [2.0.0] - 2017-12-28
 ### Changed
 - Factories use FactoryBot instead of FactoryGirl